Add a tool that create a clang compilation database from a bazel extra action.
Clang compilation databases are used by clang-based tooling to understand how to parse the source files. Bazel's extra action interface is experimental and subject to change. Currently it does not include the path to the file in the source tree, rather it provides an ephemeral path inside the sandbox. Given this restriction, any clang tooling would need to be integrated into the extra_action run, instead of building a compilation database for the whole codebase and then using that.